First Fixed Website Development Project
The design should align with the brand's overall identity and message; by aligning with the brand identity and prioritizing user experience in the design, a website can provide a positive experience. I wanted to create a website that implores that. 
The colors I picked out, dark green, light green, grey, white and black, all tie together to reflect nature and the farm itself. The slight curve of the boxes and their light grey color creates a soft touch, contributing to the brands identity. The color temperature contributes the professionalism and calmness of the website's tone. Navigating a new website can be confusing at first, so it was important the website showed consistency between the style of all links, headings, and text. 
The layout and navigation of the website should be intuitive and easy to use. I have done this by ensuring the index page includes a brief overview of all the website's content. If this was a full website each of the “read more” tabs would link to a separate page that can also be reached by links in the navigation bar. The content has been organized in a way that creates good contrast between background, text, and images. 
I used Trebuchet MS for all the headings (h1-h5 tags), the h1 tag is used for the "company's" tag line, Arial was used to style the body text. The h2 tags are bolded and call out important information on the page, whereas h3 tags are thin and used as subheadings to label specific information. The smallest size font is 10px, so a sans-serif font makes the website have increased readability. To separate div containers, I styled the 2-3 column rows with similar code, and the single-column rows were styled uniquely. 
Lastly, I used ChatGPT to generate the brand name and the body text; the tree in the logo was also created by AI. In photoshop the text, "Sweet Tree Farms," was added into the logo. This website was created using what I learned over the first three months of 2023.